requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.8.2  # https://github.com/evansd/whitenoise
redis==5.2.0  # https://github.com/redis/redis-py
hiredis==3.0.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==5.0.9  # pyup: < 5.1  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.2.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.3  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-import-export==3.0.2

tenacity==8.2.3
requests==2.32.3
beautifulsoup4==4.12.3
django-admin-extra-buttons==1.5.5
